company-student-matchmaker/requirements.txt
langchain
google-generativeai
streamlit
python-dotenv
langchain-community
langchain-google-genai



